refactor(web): extract task-header from task detail
