+++ packages/web/src/routes/teams/$teamId/projects/$projectId/tasks/$taskId.tsx
@@ -35,9 +35,9 @@ import {
 import { MarkdownProse } from '../../../../../../components/markdown-prose';
 import { MentionTextarea } from '../../../../../../components/mention-textarea';
 import { ProjectIntakeBanner } from '../../../../../../components/task-detail/project-intake-banner';
+import { TaskHeader } from '../../../../../../components/task-detail/task-header';
 import { TaskStatusBadge } from '../../../../../../components/task-status-badge';
 import { Avatar, avatarColorFromString } from '../../../../../../components/ui/avatar';
-import { Badge } from '../../../../../../components/ui/badge';
 import { Button } from '../../../../../../components/ui/button';
 import { ConfirmDialog } from '../../../../../../components/ui/confirm-dialog';
 import { InfoTooltip } from '../../../../../../components/ui/info-tooltip';
@@ -71,13 +71,6 @@ function previewCommentText(c: Comment): string {
 	if (!text) return '(non-text comment)';
 	return text.length > 60 ? `${text.slice(0, 60)}…` : text;
 }
-
-const priorityColors: Record<string, string> = {
-	urgent: 'danger',
-	high: 'warning',
-	medium: 'info',
-	low: 'neutral',
-};
 
 const EFFORT_LEVELS: { value: AgentEffort; label: string }[] = [
 	{ value: AgentEffort.Minimal, label: 'Minimal' },
@@ -354,65 +347,7 @@ function TaskDetailPage() {
 		<div className="grid grid-cols-1 lg:grid-cols-[1fr_190px] gap-5">
 			{/* Main content */}
 			<div className="min-w-0">
-				<div className="mb-1 text-[13px] font-mono text-text-muted">{task.identifier}</div>
-				<h1 className="text-xl font-medium mb-3">{task.title}</h1>
-
-				<div className="flex flex-wrap gap-1.5 mb-4">
-					<TaskStatusBadge status={task.status} />
-					<Badge color={priorityColors[task.priority] as 'neutral'}>{task.priority}</Badge>
-					{task.project_name && task.project_slug && (
-						<Link
-							to="/teams/$teamId/projects/$projectId"
-							params={{ teamId, projectId: task.project_slug }}
-							className="hover:opacity-80 transition-opacity"
-						>
-							<Badge color="info">{task.project_name}</Badge>
-						</Link>
-					)}
-					{!task.has_active_run && task.queued_wakeup && (
-						<Badge color="blue" className="gap-1" data-testid="task-queued-badge">
-							<span className="inline-block w-1.5 h-1.5 rounded-full bg-accent-blue-text" />
-							{task.queued_wakeup.blocker_identifier && task.queued_wakeup.blocker_project_slug ? (
-								<>
-									Queued behind{' '}
-									<Link
-										to="/teams/$teamId/projects/$projectId/tasks/$taskId"
-										params={{
-											teamId,
-											projectId: task.queued_wakeup.blocker_project_slug,
-											taskId: task.queued_wakeup.blocker_identifier.toLowerCase(),
-										}}
-										className="underline"
-									>
-										{task.queued_wakeup.blocker_identifier}
-									</Link>
-								</>
-							) : (
-								'Run queued'
-							)}
-						</Badge>
-					)}
-				</div>
-
-				{task.description && (
-					<div
-						className="mb-5 rounded-md border border-border bg-bg-elevated overflow-hidden"
-						data-testid="task-description-card"
-					>
-						<div className="flex items-center gap-2 px-3 py-2 border-b border-border bg-bg-muted">
-							<span className="text-xs font-medium text-text-muted">Description</span>
-						</div>
-						<div className="px-3 py-2.5">
-							<MarkdownProse
-								testId="task-description"
-								teamId={teamId}
-								projectSlug={taskProjectSlug}
-							>
-								{task.description}
-							</MarkdownProse>
-						</div>
-					</div>
-				)}
+				<TaskHeader task={task} teamId={teamId} taskProjectSlug={taskProjectSlug} />
 
 				<div
 					data-testid="pinned-progress-summary"
